Fixes to add/remove builder from project
Update the way project are unload so the project is closed that is
required to avoid issues with items holding references to the unloaded
projects.

Update remove to remove the generated files and CSharp referenced
assemblies

$(IceHome)\slice is added to the Include Directories of CSharp projects
when the builder is enabled, for C++ that is already part of the common
property sheet.